foosball: add foosball demonstration
This demo is composed by two parts: foosball table, where goals are
detected and provides a small interface with players, and a score counter.

These parts talk to each other using OIC protocol.

TODO:
    * scoreboard implementation - gtk and edison confs
    * handle goal-detector configurations on table (it only works for gtk)

For now it's possible to test it in a dev machine simulating the table
in a side and values will be printed in console by scoreboard process.

Signed-off-by: Bruno Dilly <bruno.dilly@intel.com>